super lengkap/requirements.txt
fastapi
uvicorn
qrcode
pillow
python-multipart
pydantic
mysql-connector-python
requests
